+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## \[0.1.15\] [\2020-10-15\]
 
 ### Added
 * Attribute dictionary added to class `compas_rcf.fabrication.clay_obj.ClayBullet`
+* Attribute `vkey` added to `ClayBullet` to store vertex key from `compas.datastructures.Network`
+* Simple import tests added
 
 ### Changed
 * Property `Vector` in `ClayBullet` changed to represent the bullets center as a line.